requirements.txt
numpy
matplotlib==3.3.4
opencv-python==4.5.5.64
PyYAML==5.4.1
pandas
ipython
pillow
torchmetrics
